fix(chartcuterie): Fix import tree
