+++ static/app/utils/metrics/index.spec.tsx
@@ -7,7 +7,6 @@ import {
   getDefaultAggregation,
   getFormattedMQL,
   getMetricsInterval,
-  isExtractedCustomMetric,
   isFormattedMQL,
 } from 'sentry/utils/metrics';
 import {DEFAULT_AGGREGATES} from 'sentry/utils/metrics/constants';
@@ -87,22 +86,6 @@ describe('getFormattedMQL', () => {
     });
 
     expect(result).toEqual('');
-  });
-});
-
-describe('isExtractedCustomMetric', () => {
-  it('should return true if the metric name is prefixed', () => {
-    expect(isExtractedCustomMetric({mri: 'c:custom/span_attribute_123@none'})).toBe(true);
-    expect(isExtractedCustomMetric({mri: 's:custom/span_attribute_foo@none'})).toBe(true);
-    expect(isExtractedCustomMetric({mri: 'd:custom/span_attribute_bar@none'})).toBe(true);
-    expect(isExtractedCustomMetric({mri: 'g:custom/span_attribute_baz@none'})).toBe(true);
-  });
-
-  it('should return false if the metric name is not prefixed', () => {
-    expect(isExtractedCustomMetric({mri: 'c:custom/12span_attribute_@none'})).toBe(false);
-    expect(isExtractedCustomMetric({mri: 's:custom/foo@none'})).toBe(false);
-    expect(isExtractedCustomMetric({mri: 'd:custom/_span_attribute_@none'})).toBe(false);
-    expect(isExtractedCustomMetric({mri: 'g:custom/span_attributebaz@none'})).toBe(false);
   });
 });
