+++ static/app/utils/metrics/index.tsx
@@ -332,11 +332,6 @@ export function isCustomMetric({mri}: {mri: MRI}) {
   return mri.includes(':custom/');
 }
 
-export function isExtractedCustomMetric({mri}: {mri: MRI}) {
-  // Extraced metrics are prefixed with `span_attribute_`
-  return mri.substring(1).startsWith(':custom/span_attribute_');
-}
-
 export function isVirtualMetric({mri}: {mri: MRI}) {
   return mri.startsWith('v:');
 }
